% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainRot-package.R
\docType{package}
\name{brainRot}
\alias{brainRot-package}
\alias{brainRot}
\title{brainRot: THE package for YOU.}
\description{
The \code{brainRot} package is a hybrid toolkit for:
\itemize{
\item Delivering Gen Z-inspired motivational quotes
\item Exploring fake-but-familiar indicators
\item Performing weighted estimations on external mock data
}
}
\details{
\subsection{Available Functions}{
\itemize{
\item \code{motivate_me(category)} – Get a motivational quote from a chosen category.
\item \code{simulate_brainrot(hours)}- simulate iq after a session of brainrot
\item \code{indicate_me_up()} – List all indicator names.
\item \code{what_this_mean_brah(indicator_name)} – Explain a specific indicator.
\item \code{estimate_brainrot(mockdata, variable, by = NULL)} – Calculate weighted means with \code{survey} package.
}
}

\subsection{Need Help?}{
\itemize{
\item Run \code{?function_name} to get help on any individual function - HIGHLY RECOMMENDED
\item Run \code{?brainRot} (this page) to see the full package overview.
}

The package is designed to be funny, practical and helps the user use mock data as well
}
}
\author{
\strong{Maintainer}: Atharv Deshpande \email{atharvdeshpande9@gmail.com}

}
