% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicate_me_up.R
\name{indicate_me_up}
\alias{indicate_me_up}
\title{List all available indicators}
\usage{
indicate_me_up()
}
\value{
A character vector of indicator names.
}
\description{
This function returns the names of all indicators available in the dataset.
Use this to find what columns you can explore.
}
