% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/what_this_mean_brah.R
\name{what_this_mean_brah}
\alias{what_this_mean_brah}
\title{Explain what an indicator means (brainrot edition)}
\usage{
what_this_mean_brah(indicator)
}
\arguments{
\item{indicator}{A string name of the indicator you want explained.}
}
\value{
A character string with the indicator's description.
}
\description{
Given an indicator name, this function returns a helpful (and chaotic) description.
}
