% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morivate_me.R
\name{motivate_me}
\alias{motivate_me}
\title{Return a random quote for the given category}
\usage{
motivate_me(cat)
}
\arguments{
\item{cat}{A character string matching one of the categories in the \code{quotes} dataset - grind, skibidi, npc, heartbreak, existential}
}
\value{
A single character string (one random quote). If \code{cat} doesn’t exist, throws a warning and returns \code{NA}.
}
\description{
Return a random quote for the given category
}
